+++ src/main/java/expresscarts/mixin/NewMinecartBehaviorMixin.java
@@ -1,6 +1,7 @@
 package expresscarts.mixin;
 
 import expresscarts.ExpressCarts;
+import expresscarts.ExpressCartsConfig;
 import expresscarts.ExpressMinecartEntity;
 import net.minecraft.world.entity.vehicle.AbstractMinecart;
 import net.minecraft.world.entity.vehicle.MinecartBehavior;
@@ -21,7 +22,7 @@ public abstract class NewMinecartBehaviorMixin extends MinecartBehavior {
     // If we're using this controller for an ExpressMinecartEntity, use the configured speed for that
     private void onGetMaxSpeed(CallbackInfoReturnable<Double> cir) {
         if (this.minecart instanceof ExpressMinecartEntity) {
-            cir.setReturnValue(ExpressCarts.MAX_MINECART_SPEED * (this.minecart.isInWater() ? ExpressCarts.WATER_SPEED_MULTIPLIER : 1.0) / 20.0);
+            cir.setReturnValue(ExpressCartsConfig.maxMinecartSpeed * (this.minecart.isInWater() ? ExpressCartsConfig.waterSpeedMultiplier : 1.0) / 20.0);
         }
     }
 }