+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-
-- Holding the backwards movement key while in a moving Express Minecart will now cause it to brake
+- Holding the backwards movement key while in a moving Express Minecart will now cause it to brake.
+- Values can now be configured using `/config expresscarts`.